+++ tests/integrations/dspy/cassettes/test_dspy/test_dspy_classification.yaml
@@ -55,19 +55,17 @@ interactions:
   response:
     body:
       string: !!binary |
-        H4sIAAAAAAAAAwAAAP//jFLBbqMwEL3zFZZ7DRWhyUJy25V6btXuaZMIOfZAZtfYlj1Erar8+8qQ
-        BKq2Ui9IzJv3/N7MvCWMcVR8zbg8CJKt0+mveVAvT68/1ZKeHv48Pnc/9DLzq/tj+/zb8llk2P1f
-        kHRh3UrbOg2E1gyw9CAIouq8uFsVZZkVRQ+0VoGOtMZRurBpiwbTPMsXaVak8/LMPliUEPiabRLG
-        GHvrv9GnUfDC1yybXSothCAa4OtrE2PcWx0rXISAgYQhPhtBaQ2B6a1vNuzmhgUwhC0Yij+73dY4
-        G5DwCFuzNUOLtKZGBUbCpSe7LZdTvI8PaoCnz3mouyBiZNNpfa6frv61bZy3+3DGr/UaDYZD5UEE
-        a6LXQNbxHj0ljO36OXXvonPnbeuoIvsPTBTMF6tBj4/rmaD5GSRLQk/qxXz2iV6lgATqMJk0l0Ie
-        QI3UcS2iU2gnQDJJ/dHNZ9pDcjTNd+RHQEpwBKpyHhTK94nHNg/xer9qu065N8wD+CNKqAjBx00o
-        qEWnh5vi4TUQtFWNpgHvPA6HVbsqyxZ5uS/KXPDklPwHAAD//wMAGkIixWYDAAA=
+        H4sIAAAAAAAAAwAAAP//jFLBbtswDL37KwT1GheO58JujttOuwxYgQFbEhiKRNvsZEmT6KZFkX8f
+        ZCexi3XALgbMx/f0HsnXhDGOim8Yl50g2TudflyH39nd54dShS/YfdNfX47HHw/ffx7z58dPfBUZ
+        9vAIki6sW2l7p4HQmgmWHgRBVF2XH+7LqqryYgR6q0BHWusoLWzao8E0z/Iizcp0XZ3ZnUUJgW/Y
+        NmGMsdfxG30aBc98w7LVpdJDCKIFvrk2Mca91bHCRQgYSBjiqxmU1hCY0fp2y25uWABD2IOh+LPf
+        74yzAQmfYGd2ZmqR1jSowEi49GS31d0SH+ODmuDlcx6aIYgY2Qxan+unq39tW+ftIZzxa71Bg6Gr
+        PYhgTfQayDo+oqeEsf04p+FNdO687R3VZH+BiYJ5cT/p8Xk9CzQ/g2RJ6EW9XK/e0asVkEAdFpPm
+        UsgO1Eyd1yIGhXYBJIvUf7t5T3tKjqb9H/kZkBIcgaqdB4XybeK5zUO83n+1Xac8GuYB/BNKqAnB
+        x00oaMSgp5vi4SUQ9HWDpgXvPE6H1bg6y4q8OpRVLnhySv4AAAD//wMAlr/GKGYDAAA=
     headers:
-      CF-Cache-Status:
-      - DYNAMIC
       CF-RAY:
-      - 9135167c391e2e39-BOM
+      - 913528b97e524719-BOM
       Connection:
       - keep-alive
       Content-Encoding:
@@ -75,7 +73,7 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Mon, 17 Feb 2025 10:27:58 GMT
+      - Mon, 17 Feb 2025 10:40:25 GMT
       Server:
       - cloudflare
       Transfer-Encoding:
@@ -86,10 +84,12 @@ interactions:
       - X-Request-ID
       alt-svc:
       - h3=":443"; ma=86400
+      cf-cache-status:
+      - DYNAMIC
       openai-organization:
       - wandb
       openai-processing-ms:
-      - '499'
+      - '501'
       openai-version:
       - '2020-10-01'
       strict-transport-security:
@@ -107,7 +107,7 @@ interactions:
       x-ratelimit-reset-tokens:
       - 0s
       x-request-id:
-      - req_e12b05556c98ababbb517450e8c75f28
+      - req_f91e93f061ec39e59570ec83be2c151c
     status:
       code: 200
       message: OK
