+++ tests/integrations/dspy/cassettes/test_dspy/test_dspy_lm_call.yaml
@@ -318,13 +318,13 @@ interactions:
       Cross-Origin-Resource-Policy:
       - cross-origin
       Date:
-      - Mon, 17 Feb 2025 10:27:51 GMT
+      - Mon, 17 Feb 2025 10:40:18 GMT
       ETag:
       - W/"05ebff67c6850ad022cd0cb87af139f630549dac93b850103d0f0d274d732d9c"
       Expires:
-      - Mon, 17 Feb 2025 10:32:51 GMT
+      - Mon, 17 Feb 2025 10:45:18 GMT
       Source-Age:
-      - '292'
+      - '125'
       Strict-Transport-Security:
       - max-age=31536000
       Vary:
@@ -334,19 +334,19 @@ interactions:
       X-Cache:
       - HIT
       X-Cache-Hits:
-      - '18'
+      - '14'
       X-Content-Type-Options:
       - nosniff
       X-Fastly-Request-ID:
-      - 0f250b01b8a114da71c32100f48e8d84fbd1c76c
+      - 7916558aa0e47838e44cddb24c6ddd3db56a2c01
       X-Frame-Options:
       - deny
       X-GitHub-Request-Id:
       - 150C:28ED01:6A02:C4D1:67AFC1B9
       X-Served-By:
-      - cache-bom4720-BOM
+      - cache-bom4721-BOM
       X-Timer:
-      - S1739788072.823766,VS0,VE0
+      - S1739788819.702120,VS0,VE0
       X-XSS-Protection:
       - 1; mode=block
     status:
@@ -394,17 +394,17 @@ interactions:
   response:
     body:
       string: !!binary |
-        H4sIAAAAAAAAAwAAAP//jFLLbtswELzrK7Y8W4Vlu7HjS4ECjV3kVKCHBEUhMORKYkuRNHeV5gH/
-        e0HJsRzUAXLhYWZnMLPc5wxAGC3WIFQjWbXB5l8K0t3tzfXF5uH26utGx93V98tPuy3VT8VWTJLC
-        3/1GxS+qj8q3wSIb7wZaRZSMybVYzi+Xq9V0ueiJ1mu0SVYHzhc+b40z+Ww6W+TTZV6sDurGG4Uk
-        1vAzAwB47t+U02l8EGuYTl6QFolkjWJ9HAIQ0duECElkiKVjMRlJ5R2j66P/aAyBIZDASPwBtv4v
-        KOngGwxKePQdVF3kBuPnU4+IVUcy9XCdtQd8fwxlfR2iv6MDf8Qr4ww1ZURJ3qUAxD6Int1nAL/6
-        8t2rPiJE3wYu2f9BlwyL+WAnxpWfIdmztCM+u5iccSs1sjSWTpYnlFQN6lE5blp22vgTIjvp/H+Y
-        c95Db+Pq99iPhFIYGHUZImqjXhcexyKmg3xr7LjjPrAgjPdGYckGY/oHjZXs7HAmgh6JsS0r42qM
-        IZrhVqpQFnNEvagUFiLbZ/8AAAD//wMAsFuZrjkDAAA=
+        H4sIAAAAAAAAAwAAAP//jFLBatwwEL37K6Y6r4vtbLvOXgohgeTehtBSjFYa20pkSWjGbULYfy+y
+        N+sN3UIvOrw37/HeaF4zAGG02IJQvWQ1BJtflRSe7+vrT/Xd4015eXv/cI0PneyGG/v9m1glhd89
+        ouI31Uflh2CRjXczrSJKxuRabi4uN3VdV+VEDF6jTbIucL72+WCcyauiWufFJi/rg7r3RiGJLfzI
+        AABepzfldBqfxRaK1RsyIJHsUGyPQwAiepsQIYkMsXQsVgupvGN0U/SvvSEwBBIYiT/Arf8NSjq4
+        g1kJL36EdozcY/xy6hGxHUmmHm609oDvj6Gs70L0OzrwR7w1zlDfRJTkXQpA7IOY2H0G8HMqP77r
+        I0L0Q+CG/RO6ZFhezHZiWfkZkj1Lu+DV59UZt0YjS2PpZHlCSdWjXpTLpuWojT8hspPOf4c55z33
+        Nq77H/uFUAoDo25CRG3U+8LLWMR0kP8aO+54CiwI4y+jsGGDMf2DxlaOdj4TQS/EODStcR3GEM18
+        K21oimJd1btNXUmR7bM/AAAA//8DAE9Ni/85AwAA
     headers:
       CF-RAY:
-      - 9135166919f02e39-BOM
+      - 913528a4fe354719-BOM
       Connection:
       - keep-alive
       Content-Encoding:
@@ -412,14 +412,14 @@ interactions:
       Content-Type:
       - application/json
       Date:
-      - Mon, 17 Feb 2025 10:27:55 GMT
+      - Mon, 17 Feb 2025 10:40:21 GMT
       Server:
       - cloudflare
       Set-Cookie:
-      - __cf_bm=_bbZO.AQdGsuoMdFKf.z1j8z199sT4B_FXJQZBO_OHA-1739788075-1.0.1.1-l7nrSrBE4DX1_F3o1FM2N.JjInkAQy6x9qS8CNyHL8bExloditTiYrSyqZEemdOBm5ZKpbT1vhV3Ry1oOsts7g;
-        path=/; expires=Mon, 17-Feb-25 10:57:55 GMT; domain=.api.openai.com; HttpOnly;
+      - __cf_bm=ck3XVapvaDMC0KN3gDS13MIsbEclmidogzQbjfy61ds-1739788821-1.0.1.1-D0.IuT_aTPTs_9eUmD4WiAQ8ogiyDcdwdtnO1yuMOEW8WBGmz8eWzpJpPDppBahBRvjbW1L4IE9NG8Wo_QOm8w;
+        path=/; expires=Mon, 17-Feb-25 11:10:21 GMT; domain=.api.openai.com; HttpOnly;
         Secure; SameSite=None
-      - _cfuvid=yPGJaeA9sBz8SQzdNIhSKH1bYbRLmRFApeSxx5tXt8M-1739788075272-0.0.1.1-604800000;
+      - _cfuvid=sqZvJdza_YaM0l73d0D3E4EyxMOyzgqDM5lzNEGhcOM-1739788821996-0.0.1.1-604800000;
         path=/; domain=.api.openai.com; HttpOnly; Secure; SameSite=None
       Transfer-Encoding:
       - chunked
@@ -434,7 +434,7 @@ interactions:
       openai-organization:
       - wandb
       openai-processing-ms:
-      - '357'
+      - '415'
       openai-version:
       - '2020-10-01'
       strict-transport-security:
@@ -452,7 +452,7 @@ interactions:
       x-ratelimit-reset-tokens:
       - 0s
       x-request-id:
-      - req_49477c8a1ea1887ab23adb85d4c2e7e5
+      - req_40bce2a18f4f638a509cc961222911c2
     status:
       code: 200
       message: OK
